added new primitive function type
I wanted to add a few new primitive functions, but since their types are not intFunc, something had to be rearranged in the Obj type. I added a new Prim type, somewhat analogous to the Obj type and a subtype thereof. This Prim type can be extended later to include other types of primitive functions. (All of this could have been avoided, of course, in a language other than C.)

The function lookup_prim_name in print.c was absolutely butchered in the conversion. it works, but it could stand some improvement.

-- Besides all that, I started experimenting with using even more macros. Specifically, using macros in place of ugly and frequent selecting of various pieces of Objs. The readability of llh.c will be greatly improved with these selector macros.