added restore macro to simplify restore function
the restore function requires the address of its argument. this is inelegant, but can be fixed with a macro. whereas before restore would be called like restore(&expr), it can now be called like restore(expr), which the preprocessor transforms into _restore(&expr), _restore being the new name for the actual function