added special case of simple function (SICP ex 5.32)
SICP exercise 5.32 :

a. Extend the explicit-control evaluator to recognize as a separate class of expressions combinations whose
operator is a symbol, and to take advantage of this fact in evaluating such expressions.
b. Alyssa P. Hacker suggests that by extending the evaluator to recognize more and more special cases we could
incorporate all the compiler's optimizations, and that this would eliminate the advantage of compilation altogether.
What do you think of this idea?

ec_main's new SIMPLE_FUNC label takes care of this. duplicated code from SIMPLE_FUNC and DID_FUNC has been moved to the new CHECK_NO_ARGS label.

what are the efficiency gains? two saves (and two restores) per simple function called, since the env and the arglist don't need to be saved. for instance, (+ 4 5) takes 8 saves in the original, and 6 with the new arrangement. more substantially, (recursive_fibonacci 10) goes from 7048 saves to 5526.

of course, this is still not