added SIMPLE_ARG to reduce saves
playing with the compiler has given me a sense of how the interpreter can be made more efficient. one way is by not saving the env and cont registers in the case of simple arguments, as it was in the case of simple functions. The code here is a little trickier, but I think I've done a good job at keeping it comprehensible.

number for (recursive_fibonacci_interpreted 10):

BEFORE:
Total number of saves: 5526
Maximum stack depth: 56
Total times syntax checked: 11316

AFTER:
Total number of saves: 3541
Maximum stack depth: 53
Total times syntax checked: 11316