extending extendEnv for variable-arity functions
after banging my head against a wall for weeks, this turned out to be a shockingly easy feature to implement. nothing had to be touched in ec_main or llh, just a slight change in extendEnv.

now we can say (define list (lambda s s)). but to use this for more complicated functions, we have to implement list operations as real list operations, not church simulations