Use print macros
The print code previously featured this comment:

/* we could use macros to make these more concise,
but this way the code has a wysiwyg feel */

With the recent formatting standardization, it's no longer possible to
provide that "wysiwyg feel" (if that ever worked to begin with), so
switch to using print macros.

That said, this seems like a strange approach. Is there a more
standard way to do this kind of thing?